moved item tests from ordercontroller to itemcontroller
